Move fixtures to a public module
